SDIT-3171 Sync sexual orientation from nomis

+++ src/test/kotlin/uk/gov/justice/digital/hmpps/nomisprisonerapi/helper/builders/OffenderProfileDetail.kt
@@ -33,11 +33,11 @@ class OffenderProfileDetailBuilder(
 ) : OffenderProfileDetailDsl {
 
   fun build(
-    listSequence: Long,
+    listSequence: Int,
     profileTypeId: String,
     profileCodeId: String?,
     offenderBooking: OffenderBooking,
-    sequence: Long,
+    sequence: Int,
     profile: OffenderProfile? = null,
   ): OffenderProfileDetail = OffenderProfileDetail(
     id = OffenderProfileDetailId(
